Pulled all of the methods relating to Application Sections out of `ApplicationController` to their own `AppSectionController`
