+++ src/main/scala/rifs/business/controllers/ApplicationController.scala
@@ -12,10 +12,9 @@ import rifs.business.Config
 import rifs.business.data.{ApplicationFormOps, ApplicationOps, OpportunityOps}
 import rifs.business.models.{ApplicationFormId, ApplicationId}
 import rifs.business.notifications.NotificationService
-import rifs.business.restmodels.{ApplicationDetail, ApplicationSectionDetail}
+import rifs.business.restmodels.ApplicationDetail
 
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Random
 
 class ApplicationController @Inject()(applications: ApplicationOps,
                                       appForms: ApplicationFormOps,
@@ -63,50 +62,6 @@ class ApplicationController @Inject()(applications: ApplicationOps,
   def deleteAll() = Action.async { implicit request =>
     applications.deleteAll.map(_ => NoContent)
   }
-
-  def section(id: ApplicationId, sectionNumber: Int) =
-    Action.async(applications.fetchSection(id, sectionNumber).map(jsonResult(_)))
-
-  def sectionDetail(id: ApplicationId, sectionNumber: Int) = Action.async {
-    val ft = for {
-      a <- OptionT(applications.application(id))
-      f <- OptionT(appForms.byId(a.applicationFormId))
-      o <- OptionT(opps.opportunity(f.opportunityId))
-      fs <- OptionT.fromOption(f.sections.find(_.sectionNumber == sectionNumber))
-    } yield {
-      ApplicationSectionDetail(
-        a.id,
-        f.sections.length,
-        a.sections.count(_.completedAt.isDefined),
-        o.summary,
-        fs,
-        a.sections.find(_.sectionNumber == sectionNumber))
-    }
-
-    ft.value.map(jsonResult(_))
-  }
-
-  def sections(id: ApplicationId) =
-    Action.async(applications.fetchSections(id).map(os => Ok(Json.toJson(os))))
-
-  def saveSection(id: ApplicationId, sectionNumber: Int) = Action.async(parse.json[JsObject]) { implicit request =>
-    applications.saveSection(id, sectionNumber, request.body).map(_ => NoContent)
-  }
-
-  def clearSectionCompletedDate(id: ApplicationId, sectionNumber: Int) = Action.async { implicit request =>
-    applications.clearSectionCompletedDate(id, sectionNumber).map { count =>
-      if (count > 0) NoContent else NotFound
-    }
-  }
-
-  def completeSection(id: ApplicationId, sectionNumber: Int) = Action.async(parse.json[JsObject]) { implicit request =>
-    applications.saveSection(id, sectionNumber, request.body, Some(DateTime.now(DateTimeZone.UTC))).map(_ => NoContent)
-  }
-
-  def deleteSection(id: ApplicationId, sectionNumber: Int) = Action.async { implicit request =>
-    applications.deleteSection(id, sectionNumber).map(_ => NoContent)
-  }
-
 
   def submit(id: ApplicationId) = Action.async { _ =>
     import Config.config.rifs.{email => emailConfig}
